+++ packages/fireworks/src/fireworks-embedding-settings.ts
@@ -1,0 +1,9 @@
+import { OpenAICompatibleEmbeddingSettings } from '@ai-sdk/openai-compatible';
+
+// Below is just a subset of the available models.
+export type FireworksEmbeddingModelId =
+  | 'nomic-ai/nomic-embed-text-v1.5'
+  | (string & {});
+
+export interface FireworksEmbeddingSettings
+  extends OpenAICompatibleEmbeddingSettings {}