+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @ai-sdk/openai-compatible
 
+## 0.0.13
+
+### Patch Changes
+
+- 6564812: feat (provider/openai-compatible): Add'l exports for customization.
+
+## 0.0.12
+
+### Patch Changes
+
+- 70003b8: feat (provider/openai-compatible): Allow extending messages via metadata.
+
+## 0.0.11
+
+### Patch Changes
+
+- 5ed5e45: chore (config): Use ts-library.json tsconfig for no-UI libs.
+- 307c247: fix (provider/openai-compatible): Fix docs link to more info.
+- Updated dependencies [5ed5e45]
+  - @ai-sdk/provider-utils@2.0.5
+  - @ai-sdk/provider@1.0.3
+
+## 0.0.10
+
+### Patch Changes
+
+- baae8f4: feat (provider/deepinfra): Add DeepInfra provider.
+
+## 0.0.9
+
+### Patch Changes
+
+- 9c7653b: feat (docs): Update OpenAI Compatible docs for new package.
+
+## 0.0.8
+
+### Patch Changes
+
+- 6faab13: feat (provider/openai-compatible): simulated streaming setting
+
+## 0.0.7
+
+### Patch Changes
+
+- ad2bf11: feat (provider/fireworks): Add Fireworks provider.
+
 ## 0.0.6
 
 ### Patch Changes