+++ packages/togetherai/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @ai-sdk/togetherai
 
+## 0.0.15
+
+### Patch Changes
+
+- ac101a2: feat (examples): Use shared base e2e feature test suite.
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [6564812]
+  - @ai-sdk/openai-compatible@0.0.13
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [70003b8]
+  - @ai-sdk/openai-compatible@0.0.12
+
+## 0.0.12
+
+### Patch Changes
+
+- 5ed5e45: chore (config): Use ts-library.json tsconfig for no-UI libs.
+- 8974d6d: chore (provider/togetherai): Incorporate the latest model ids.
+- Updated dependencies [5ed5e45]
+- Updated dependencies [307c247]
+  - @ai-sdk/openai-compatible@0.0.11
+  - @ai-sdk/provider-utils@2.0.5
+  - @ai-sdk/provider@1.0.3
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [baae8f4]
+  - @ai-sdk/openai-compatible@0.0.10
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [9c7653b]
+  - @ai-sdk/openai-compatible@0.0.9
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [6faab13]
+  - @ai-sdk/openai-compatible@0.0.8
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [ad2bf11]
+  - @ai-sdk/openai-compatible@0.0.7
+
 ## 0.0.7
 
 ### Patch Changes