+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @ai-sdk/xai
 
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [6564812]
+  - @ai-sdk/openai-compatible@0.0.13
+
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [70003b8]
+  - @ai-sdk/openai-compatible@0.0.12
+
+## 1.0.12
+
+### Patch Changes
+
+- 5ed5e45: chore (config): Use ts-library.json tsconfig for no-UI libs.
+- Updated dependencies [5ed5e45]
+- Updated dependencies [307c247]
+  - @ai-sdk/openai-compatible@0.0.11
+  - @ai-sdk/provider-utils@2.0.5
+  - @ai-sdk/provider@1.0.3
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [baae8f4]
+  - @ai-sdk/openai-compatible@0.0.10
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [9c7653b]
+  - @ai-sdk/openai-compatible@0.0.9
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [6faab13]
+  - @ai-sdk/openai-compatible@0.0.8
+
+## 1.0.8
+
+### Patch Changes
+
+- 50821de: feat (docs): Use new grok-2 model in xai example code.
+
+## 1.0.7
+
+### Patch Changes
+
+- 4e9032c: feat (provider/xai): Add grok-2 models, use openai-compatible base impl.
+
 ## 1.0.6
 
 ### Patch Changes